Working on full width article layout. (in progress)
Most of the new layout is completed.  I will be adding a few extra
styles for images and other in line elements.

Responsive design has not yet be done for these pages but will be
completed once all text and images style have been completed.

This shouldn't have any knock on effect on the rest of the site, except
for the background color changed to white.